requirements.txt

cryptography
pycryptodome
matplotlib
